requirements.txt
fastapi
uvicorn
requests
psycopg2-binary
pydantic
python-dotenv
prometheus-client
python-json-logger
sentence-transformers
pymilvus
redis
pandas
matplotlib
scikit-learn
pyyaml
numpy